fix: fix component
