+++ src/App.jsx
@@ -1,6 +1,7 @@
 import "./App.css";
-import { NavBar, HeroSection, Footer } from "./components/LandingPage";
-import { ArticleSection } from "./components/ArticleSection";
+import { NavBar, HeroSection } from "./components/LandingPage";
+import ArticleSection from "./components/ArticleSection";
+import Footer from "./components/Footer";
 
 function App() {
   return (
